+++ hmc/coupler/cpl_hmc_builder.py
@@ -45,6 +45,7 @@ class ModelBuilder:
             template_time=self.obj_args.obj_template_time_ref,
             template_run_def=self.obj_run.obj_template_run_filled,
             template_run_ref=self.obj_args.obj_template_run_ref,
+            template_analysis_def=self.obj_run.obj_template_analysis_filled,
             template_static=self.obj_args.obj_template_dset_static_ref,
             template_dynamic=self.obj_args.obj_template_dset_dynamic_ref)