+++ hmc/driver/dataset/drv_dataset_hmc_io_dynamic_outcome.py
@@ -22,6 +22,8 @@ from copy import deepcopy
 from hmc.algorithm.io.lib_data_io_generic import create_darray_3d, create_darray_2d, write_dset, create_dset
 from hmc.algorithm.io.lib_data_zip_gzip import zip_filename
 
+from hmc.algorithm.utils.lib_utils_analysis import compute_domain_mean, \
+    compute_catchment_mean_serial, compute_catchment_mean_parallel_sync, compute_catchment_mean_parallel_async
 from hmc.algorithm.utils.lib_utils_system import split_path, create_folder, copy_file
 from hmc.algorithm.utils.lib_utils_string import fill_tags2string
 from hmc.algorithm.utils.lib_utils_list import flat_list
@@ -50,7 +52,7 @@ class DSetManager:
                  dset_list_format=None,
                  dset_list_type=None,
                  dset_list_group=None,
-                 template_time=None,
+                 template_time=None, template_analysis_def=None,
                  model_tag='hmc', datasets_tag='datasets',
                  coord_name_geo_x='Longitude', coord_name_geo_y='Latitude', coord_name_time='time',
                  dim_name_geo_x='west_east', dim_name_geo_y='south_north', dim_name_time='time',
@@ -211,6 +213,39 @@ class DSetManager:
 
         self.column_sep = ';'
 
+        self.template_analysis_def = template_analysis_def
+        if self.template_analysis_def is not None:
+
+            self.list_variable_selected = ['SM']
+            self.tag_variable_fields = '{var_name}:hmc_outcome_datasets:{domain_name}'
+
+            self.flag_analysis_ts_domain = True
+
+            if 'analysis_catchment' in list(self.template_analysis_def.keys()):
+                self.flag_analysis_ts_catchment = self.template_analysis_def['analysis_catchment']
+            else:
+                self.flag_analysis_ts_catchment = False
+
+            if 'analysis_mp' in list(self.template_analysis_def.keys()):
+                self.flag_analysis_ts_catchment_mode = self.template_analysis_def['analysis_mp']
+            else:
+                self.flag_analysis_ts_catchment_mode = False
+
+            if 'analysis_cpu' in list(self.template_analysis_def.keys()):
+                self.flag_analysis_ts_catchment_cpu = self.template_analysis_def['analysis_cpu']
+            else:
+                self.flag_analysis_ts_catchment_cpu = 1
+
+        else:
+            self.list_variable_selected = ['SM']
+            self.tag_variable_fields = '{var_name}:hmc_outcome_datasets:{domain_name}'
+
+            self.flag_analysis_ts_domain = True
+            self.flag_analysis_ts_catchment = False
+
+            self.flag_analysis_ts_catchment_mode = False
+            self.flag_analysis_ts_catchment_cpu = 1
+
     def copy_data(self, dset_model_dyn, dset_destination_dyn, columns_excluded=None):
 
         # Starting info
@@ -316,7 +351,7 @@ class DSetManager:
                                 copy_file(file_path_tmp_step, file_path_destination_step)
                             else:
                                 log_stream.warning(' ===> Copy file for variable "' + var_destination_step + '"'
-                                                   ' FAILED. File "' + file_path_tmp_step + '" does not exist!')
+                                                   ' SKIPPED. File "' + file_path_tmp_step + '" does not exist!')
                     else:
                         log_stream.warning(' ===> Copy file for variable "' + var_destination_step + '"'
                                            ' FAILED. ALL files do not exist!')
@@ -368,7 +403,7 @@ class DSetManager:
 
                     if dset_check:
 
-                        if dset_var_step != 'terrain':
+                        if dset_var_step not in ['terrain', 'mask']:
 
                             if dset_var_step not in list(dset_expected.columns):
                                 values_nan = np.zeros([dset_expected.index.__len__()])
@@ -519,12 +554,20 @@ class DSetManager:
             # Ending info
             log_stream.info(' -------> Zip data ... SKIPPED. Zip not activated')
 
-    def organize_data(self, dset_time, dset_source, dset_variable_selected='ALL'):
+    def organize_data(self, dset_time, dset_source, dset_static=None, dset_variable_selected='ALL'):
 
         # Get variable(s)
         dset_vars = dset_source[self.datasets_tag]
         # Get terrain reference
         da_terrain = self.da_terrain
+
+        if dset_static is not None:
+            if 'mask_name_list' in list(dset_static.keys()):
+                mask_name_obj = dset_static['mask_name_list']
+            else:
+                mask_name_obj = None
+        else:
+            mask_name_obj = None
 
         if dset_vars is not None:
             if (self.coord_name_geo_x in list(dset_source.keys())) and (self.coord_name_geo_x in list(dset_source.keys())):
@@ -722,19 +765,11 @@ class DSetManager:
                                                              var_geo_name='terrain', var_geo_values=self.terrain_values,
                                                              var_geo_x=self.terrain_geo_x, var_geo_y=self.terrain_geo_y,
                                                              var_geo_attrs=None)
-
-                            # Compute average values
-                            var_dset_ts_step = var_dset_grid_step.mean(dim=['south_north', 'west_east'])
-
+                            # Organize data in merged datasets
                             if var_dset_out is None:
                                 var_dset_out = var_dset_grid_step
                             else:
                                 var_dset_out = var_dset_out.merge(var_dset_grid_step, join='right')
-
-                            if var_dset_collections is None:
-                                var_dset_collections = var_dset_ts_step
-                            else:
-                                var_dset_collections = var_dset_collections.merge(var_dset_ts_step, join='right')
 
                         else:
                             log_stream.info(' --------> Organize ' + var_name_step +
@@ -746,6 +781,65 @@ class DSetManager:
                     var_dset_out = None
                     var_dset_collections = None
                     log_stream.info(' -------> Organize gridded datasets ... SKIPPED. Empty selected variables.')
+
+                # Time-Series Analysis
+                log_stream.info(' --------> Compute time-series analysis over domain ... ')
+                if self.flag_analysis_ts_domain:
+                    var_dset_ts_domain = compute_domain_mean(
+                        var_dset_out, tag_variable_fields=self.tag_variable_fields,
+                        template_variable_domain='DomainAverage')
+
+                    if var_dset_collections is None:
+                        var_dset_collections = var_dset_ts_domain
+                    else:
+                        var_dset_collections = var_dset_collections.merge(var_dset_ts_domain, join='right')
+
+                    log_stream.info(' --------> Compute time-series analysis over domain ... DONE')
+                else:
+                    log_stream.info(' --------> Compute time-series analysis over domain ... SKIPPED. '
+                                    'Analysis not activated')
+
+                log_stream.info(' --------> Compute time-series analysis over catchments ... ')
+                if self.flag_analysis_ts_catchment:
+
+                    if mask_name_obj is not None:
+
+                        if not self.flag_analysis_ts_catchment_mode:
+                            var_dset_ts_catchment = compute_catchment_mean_serial(
+                                var_dset_out, mask_name_obj,
+                                variable_domain_fields=self.tag_variable_fields,
+                                variable_selected_list=self.list_variable_selected)
+                        elif self.flag_analysis_ts_catchment_mode:
+                            # var_dset_ts_catchment = compute_catchment_mean_parallel_sync(
+                            #    var_dset_out, mask_name_obj, cpu_n=self.flag_analysis_ts_catchment_cpu,
+                            #    variable_domain_fields = self.tag_variable_fields,
+                            #    variable_selected_list=self.list_variable_selected)
+                            var_dset_ts_catchment = compute_catchment_mean_parallel_async(
+                                var_dset_out, mask_name_obj,
+                                cpu_n=self.flag_analysis_ts_catchment_cpu,
+                                variable_domain_fields=self.tag_variable_fields,
+                                variable_selected_list=self.list_variable_selected)
+                        else:
+                            log_stream.error(' ===> Catchments analysis mode not allowed')
+                            raise RuntimeError('Unexpected catchments analysis condition')
+
+                        if var_dset_collections is None:
+                            var_dset_collections = var_dset_ts_catchment
+                        else:
+                            var_dset_collections = var_dset_collections.merge(var_dset_ts_catchment, join='right')
+
+                        log_stream.info(' --------> Compute time-series analysis over catchments ... DONE')
+
+                    else:
+                        log_stream.info(' --------> Compute time-series analysis over catchments ... SKIPPED.'
+                                        ' Catchment mask obj not defined')
+                else:
+                    log_stream.info(' --------> Compute time-series analysis over catchments ... SKIPPED.'
+                                    ' Analysis not activated')
+
+                log_stream.info(' -------> Organize gridded datasets ... DONE')
+
+
 
             else: